requirements/base.txt
Django==4.2.*
djangorestframework>=3.14,<3.15
djangorestframework-simplejwt>=5.3,<6.0
dj-rest-auth>=7.0,<8.0
django-allauth>=65.0,<66.0
django-cors-headers>=4.3,<5.0
django-environ>=0.11,<1.0
django-simple-history>=3.4,<4.0
psycopg2-binary>=2.9,<3.0
stripe>=7.0,<9.0
Pillow>=10.0,<11.0
whitenoise>=6.6,<7.0
